feat: overhauled hook system
The previous hook system was clumsy and unnecessarily complex. Using a function to return the hooks was useless, since we may as well just register a hook and then - if we decide not to use it - return undefined.

It was also difficult to access hooks, because plugins had to manually parse and resolve hooks from the raw hooks map. Now hooks are provided not as a map, but as a handy-dandy accessor function returning a single promise.

BREAKING CHANGE: plugin.hooks may no longer be a function
BREAKING CHANGE: hooks is no longer provided as an object, but as a function